Wifi code with randomizer
With this code 5 boards can be connected to the wifi where they all have a different number (by changing the number on each board), it this code there is a list of which board is on and off. Because we want 2 poles to do something if a person is connecting with one pole, there is a randomizer in the code, this means that if a pole is false in the attendence list there is a 50 procent change of the pole going on or staying off.