Create Wifi code for board 1
This is the code for the first board, here all the boards will listen to each other aswell as to them selves. If nobody is near the boards, all the different boards will turn on there neopixels and do the chase animation. If one board sees somebody with one of the sensors that board will turn there Neopixels on and 2 others (the highest board number and the lowest) let their neopixels do the chase animation and the others will be off. When there are two boards that see somebody, those boards turn there neopixels on, and also a third board (the one with the lowest number) will turn there neopixels on.