Create Final prototype code
This is the code we will use for the final prototype. All the boards listen to each other aswell as to them selves. When uploading to the boards we changed the number of the board (line 22 and 23) and the interval to start the neopixel animation depending on the boardnr (line 90). If nobody is near the boards, all the different boards will turn on there neopixels and do the chase animation. If one board sees somebody with one of the sensors that board will turn there Neopixels on and 2 others (the highest board number and the lowest) let their neopixels do the chase animation and the others will be off. When there are two boards that see somebody, those boards turn there neopixels on, and also a third board (the one with the lowest number) will turn there neopixels on.